requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pymupdf==1.24.1
easyocr==1.7.0
transformers==4.36.2
torch==2.1.2
pillow==10.1.0
numpy==1.26.3
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.2
aiofiles==23.2.1
requests==2.31.0
spacy==3.7.2
sentence-transformers==3.0.1
faiss-cpu==1.7.4
redis==5.0.1
slowapi==0.1.9
pyjwt==2.10.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
prometheus-client==0.19.0
structlog==24.1.0
colorlog==6.8.0
python-json-logger==2.0.7
